Parse longs fetched from Datastore to int
When voteCount is fetched from Datastore it is converted to a Long (Datastore stores integers as 64-bit integers, so they are converted to a Long on retrieval), but for our purposes, voteCount should be of int type, so converting to an integer when fetched.

+++ GoMeet/src/main/java/com/google/sps/servlets/VoteMeetingTimeServlet.java
@@ -83,7 +83,7 @@ public class VoteMeetingTimeServlet extends HttpServlet {
       return;
     }
 
-    Long currentVotes = (long) meetingTime.getProperty(MeetingTimeFields.VOTE_COUNT);
+    int currentVotes = ((Long) meetingTime.getProperty(MeetingTimeFields.VOTE_COUNT)).intValue();
     meetingTime.setProperty(MeetingTimeFields.VOTE_COUNT, currentVotes + 1);
     meetingTime.setProperty(MeetingTimeFields.VOTERS, voters);
     datastore.put(meetingTime);
